+++ CHANGELOG.md
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 #### edges.py
 - BSplineCurve3D: move_frame_along
-- arc2d: start and end angle, and arc angle.
+- Arc2D: start and end angle, and arc angle.
 
 #### faces.py
 - Toroidalface ConicalFace intersections.